Data.Array.Accelerate.System.Random.MWC

Description

Random number generation backed by MWC.

Generating random arrays

type (:~>) sh e = sh -> GenIO -> IO e

A PRNG from indices to variates

uniform :: (Shape sh, Elt e, Variate e) => sh :~> e

Uniformly distributed random variates.

uniformR :: (Shape sh, Elt e, Variate e) => (e, e) -> sh :~> e

Uniformly distributed random variates in a given range.

randomArray :: (Shape sh, Elt e) => (sh :~> e) -> sh -> IO (Array sh e)

Generate an array of random values. The generator for variates is seeded from the system's fast source of pseudo-random numbers (see: createSystemRandom)

randomArrayWith :: (Shape sh, Elt e) => GenIO -> (sh :~> e) -> sh -> IO (Array sh e)

Generate an array of random values using the supplied generator.

Re-export MWC-Random